MemoryPack序列化Dictionary<int...，key为int时会有未知4字节且其补位在各平台不一致，改用long作为key。

+++ Unity/Assets/Scripts/Model/Share/LockStep/Battle/Buff/BuffComponent.cs
@@ -10,6 +10,6 @@ namespace ET
     public partial class BuffComponent : LSEntity, IAwake, IDestroy, ILSUpdate, ISerializeToEntity
     {
         [BsonDictionaryOptions(DictionaryRepresentation.ArrayOfArrays)]
-        public SortedDictionary<int, long> IdBuffMap = new();
+        public SortedDictionary<long, long> IdBuffMap = new();
     }
 }
